Share plugin migration test helpers
