+++ tests/test_smart_field_migration.py
@@ -20,10 +20,16 @@ along with Smart Notes.  If not, see <https://www.gnu.org/licenses/>.
 import json
 from copy import deepcopy
 from pathlib import Path
-from typing import Any, Optional, cast
 
 import pytest
-from anki.decks import DeckId
+from support import (
+    DECK_ID,
+    NOTE_TYPE_ID,
+    install_fake_anki,
+    install_fake_sentry,
+    install_migration_alert,
+    use_temp_sqlite,
+)
 
 from src.database.legacy_config_migration import migrate_legacy_config_to_database
 from src.database.migrations import apply_database_migrations
@@ -31,51 +37,10 @@ from src.models import DEFAULT_EXTRAS, FieldExtras
 from src.models.smart_fields import ChatSmartFieldSettings
 from src.services.smart_field_service import SmartFieldService
 
-NOTE_TYPE_ID = 123
-DECK_ID = cast(DeckId, 1)
-
-
-class FakeAddonManager:
-    def __init__(self, addon_config: dict[str, Any]) -> None:
-        self.addon_config = addon_config
-        self.written_config: Optional[dict[str, Any]] = None
-
-    def getConfig(self, addon_name: str) -> dict[str, Any]:
-        return self.addon_config
-
-    def writeConfig(self, addon_name: str, addon_config: dict[str, Any]) -> None:
-        self.written_config = deepcopy(addon_config)
-
-
-class FakeConfig:
-    def __init__(self, addon_config: dict[str, Any]) -> None:
-        self.addon_config = addon_config
-
-    @property
-    def did_migrate_smart_fields_to_sqlite(self) -> bool:
-        return bool(self.addon_config.get("did_migrate_smart_fields_to_sqlite"))
-
-    def __getattr__(self, key: str) -> object:
-        return self.addon_config.get(key)
-
-
-class FakeSentry:
-    def __init__(self) -> None:
-        self.captured: list[Exception] = []
-
-    def capture_exception(self, error: Exception) -> None:
-        self.captured.append(error)
-
 
 @pytest.fixture(autouse=True)
 def sqlite_database(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    import src.database.connection
-
-    monkeypatch.setattr(
-        src.database.connection,
-        "get_database_path",
-        lambda: str(tmp_path / "smart_notes.sqlite3"),
-    )
+    use_temp_sqlite(monkeypatch, tmp_path)
     apply_database_migrations()
 
 
@@ -232,62 +197,6 @@ def test_migrate_legacy_smart_field_config_does_nothing_after_successful_migrati
 
     assert fake_mw.addonManager.written_config is None
     assert not (tmp_path / "user_files").exists()
-
-
-def install_fake_anki(
-    monkeypatch: pytest.MonkeyPatch,
-    addon_config: dict[str, Any],
-    tmp_path: Path,
-) -> Any:
-    import src.database.legacy_config_migration
-    import src.smart_field_prompt_map
-
-    class FakeModels:
-        def by_name(self, note_type: str) -> Optional[dict[str, Any]]:
-            if note_type != "Basic":
-                return None
-            return {"id": NOTE_TYPE_ID}
-
-    class FakeCollection:
-        models = FakeModels()
-
-    class FakeMw:
-        def __init__(self) -> None:
-            self.addonManager = FakeAddonManager(addon_config)
-            self.col = FakeCollection()
-
-    fake_mw = FakeMw()
-    monkeypatch.setattr(src.database.legacy_config_migration, "mw", fake_mw)
-    monkeypatch.setattr(src.smart_field_prompt_map, "mw", fake_mw)
-    monkeypatch.setattr(
-        src.database.legacy_config_migration, "config", FakeConfig(addon_config)
-    )
-    monkeypatch.setattr(
-        src.database.legacy_config_migration,
-        "get_user_files_path",
-        lambda filename: str(tmp_path / "user_files" / filename),
-    )
-    return fake_mw
-
-
-def install_migration_alert(
-    monkeypatch: pytest.MonkeyPatch, messages: list[tuple[object, ...]]
-) -> None:
-    import src.database.legacy_config_migration
-
-    monkeypatch.setattr(
-        src.database.legacy_config_migration,
-        "show_message_box",
-        lambda *args: messages.append(args),
-    )
-
-
-def install_fake_sentry(monkeypatch: pytest.MonkeyPatch) -> FakeSentry:
-    import src.sentry
-
-    fake_sentry = FakeSentry()
-    monkeypatch.setattr(src.sentry, "sentry", fake_sentry)
-    return fake_sentry
 
 
 def chat_extras() -> FieldExtras:
